Eliminate self.security from API tests (breaks abstractions).
