+++ lib/galaxy_test/api/test_configuration.py
@@ -2,9 +2,6 @@ from galaxy_test.base.api_asserts import (
     assert_has_keys,
     assert_not_has_keys,
 )
-from galaxy_test.base.populators import (
-    LibraryPopulator
-)
 from ._framework import ApiTestCase
 
 TEST_KEYS_FOR_ALL_USERS = [
@@ -26,10 +23,6 @@ TEST_KEYS_FOR_ADMIN_ONLY = [
 
 class ConfigurationApiTestCase(ApiTestCase):
 
-    def setUp(self):
-        super(ConfigurationApiTestCase, self).setUp()
-        self.library_populator = LibraryPopulator(self.galaxy_interactor)
-
     def test_normal_user_configuration(self):
         config = self._get_configuration()
         assert_has_keys(config, *TEST_KEYS_FOR_ALL_USERS)
@@ -39,18 +32,6 @@ class ConfigurationApiTestCase(ApiTestCase):
         config = self._get_configuration(admin=True)
         assert_has_keys(config, *TEST_KEYS_FOR_ALL_USERS)
         assert_has_keys(config, *TEST_KEYS_FOR_ADMIN_ONLY)
-
-    def test_admin_decode_id(self):
-        new_lib = self.library_populator.new_library('DecodeTestLibrary')
-        decode_response = self._get("configuration/decode/" + new_lib["id"], admin=True)
-        response_id = decode_response.json()["decoded_id"]
-        decoded_library_id = self.security.decode_id(new_lib["id"])
-        assert decoded_library_id == response_id
-        # fake valid folder id by prepending F
-        valid_encoded_folder_id = 'F' + new_lib["id"]
-        folder_decode_response = self._get("configuration/decode/" + valid_encoded_folder_id, admin=True)
-        folder_response_id = folder_decode_response.json()["decoded_id"]
-        assert decoded_library_id == folder_response_id
 
     def test_normal_user_decode_id(self):
         decode_response = self._get("configuration/decode/badhombre", admin=False)
